+++ templates/home.html
@@ -24,7 +24,7 @@
  <h1 class="my-2">قائمة الأدوات</h1>
     <ul class="nav flex-column">
         <li class="nav-item">
-            <a  class="nav-link" href="{{ url_for('words_count_tool') }}">1 ) حساب عدد الكلمات</a>
+            <a class="nav-link" href="{{ url_for('words_count_page.words_count_tool') }}">1 ) حساب عدد الكلمات</a>
         </li>
     </ul>
 </body>
